Move new models to Report APIs
